refactor(occlusion): drop occlusionDilate in favour of occlusionContourPx
Two overlapping claim-shaping options shipped in the same unreleased change.
`occlusionDilate` was the first attempt and grows the claim in whole id-map
cells, so a 2-cell pad became a ~20px square-kernel slab; `occlusionContourPx`
was built to fix exactly that, rastering the id-map at 4x internal resolution
for coverage-aware claims with an elliptical margin in screen px.

Contour covers the case dilate existed for — measured, a px margin reaches a
full output cell of clean ground on every side at supersample 1 and 2 — and
its isotropic screen reach is the better shape besides: dilate's
8-neighbourhood grew twice as far vertically as horizontally on a 8x16 cell.
The one narrowing is reach, not precision: contour's stamp clamps at
6/supersample output cells where dilate reached 8.

Neither option was ever released, so this is a clean removal with no shim.
