+++ packages/vue/src/glyphcss/scene/GlyphMesh.occlusion.test.ts
@@ -65,7 +65,6 @@ describe("GlyphMesh (Vue) — per-mesh ramp / ambient / occlusion props", () => 
       glyphPalette: "dense",
       ambientIntensity: 0.75,
       occlusionPriority: 1,
-      occlusionDilate: 2,
       occlusionClaim: "geometry",
       occlusionContourPx: 3,
     });
@@ -73,7 +72,6 @@ describe("GlyphMesh (Vue) — per-mesh ramp / ambient / occlusion props", () => 
       glyphPalette: "dense",
       ambientIntensity: 0.75,
       occlusionPriority: 1,
-      occlusionDilate: 2,
       occlusionClaim: "geometry",
       occlusionContourPx: 3,
     });
@@ -85,7 +83,6 @@ describe("GlyphMesh (Vue) — per-mesh ramp / ambient / occlusion props", () => 
     expect(t.glyphPalette).toBeUndefined();
     expect(t.ambientIntensity).toBeUndefined();
     expect(t.occlusionPriority).toBeUndefined();
-    expect(t.occlusionDilate).toBeUndefined();
     expect(t.occlusionClaim).toBeUndefined();
     expect(t.occlusionContourPx).toBeUndefined();
   });
