+++ packages/vue/src/glyphcss/scene/GlyphMesh.ts
@@ -59,11 +59,9 @@ export interface GlyphMeshProps {
   ambientIntensity?: number;
   /** Cross-layer occlusion class for an opaque detail mesh — a higher class claims id-map cells regardless of depth. Default 0. */
   occlusionPriority?: number;
-  /** Grow this opaque detail mesh's id-map claim by N output cells (clamped to 8). Never steals from another detail mesh. Default 0. */
-  occlusionDilate?: number;
   /** Claim shape: `"alpha"` (default, texel-opaque cells only) or `"geometry"` (full triangle footprint). */
   occlusionClaim?: "alpha" | "geometry";
-  /** Coverage-aware contour claim with a screen-px margin around this mesh's ink. Omitted = point-sampled claims. */
+  /** Coverage-aware contour claim with a screen-px margin of clean ground around this mesh's ink. Never steals from another detail mesh. Omitted = point-sampled claims. */
   occlusionContourPx?: number;
   class?: string;
   // Pointer/mouse interaction — type surface matches voxcss PolyMesh.
@@ -100,7 +98,6 @@ export const GlyphMesh = defineComponent({
     glyphPalette: { type: String, default: undefined },
     ambientIntensity: { type: Number, default: undefined },
     occlusionPriority: { type: Number, default: undefined },
-    occlusionDilate: { type: Number, default: undefined },
     occlusionClaim: { type: String as PropType<"alpha" | "geometry">, default: undefined },
     occlusionContourPx: { type: Number, default: undefined },
     class: { type: String, default: undefined },
@@ -165,7 +162,6 @@ export const GlyphMesh = defineComponent({
       if (props.glyphPalette !== undefined) t.glyphPalette = props.glyphPalette;
       if (props.ambientIntensity !== undefined) t.ambientIntensity = props.ambientIntensity;
       if (props.occlusionPriority !== undefined) t.occlusionPriority = props.occlusionPriority;
-      if (props.occlusionDilate !== undefined) t.occlusionDilate = props.occlusionDilate;
       if (props.occlusionClaim !== undefined) t.occlusionClaim = props.occlusionClaim;
       if (props.occlusionContourPx !== undefined) t.occlusionContourPx = props.occlusionContourPx;
       return t;
@@ -203,7 +199,7 @@ export const GlyphMesh = defineComponent({
 
     // Update transform on id/position/scale/rotation/castShadow/receiveShadow changes
     watch(
-      () => ({ id: props.id, position: props.position, scale: props.scale, rotation: props.rotation, castShadow: props.castShadow, receiveShadow: props.receiveShadow, density: props.density, fontSize: props.fontSize, lineHeight: props.lineHeight, transparent: props.transparent, glyphPalette: props.glyphPalette, ambientIntensity: props.ambientIntensity, occlusionPriority: props.occlusionPriority, occlusionDilate: props.occlusionDilate, occlusionClaim: props.occlusionClaim, occlusionContourPx: props.occlusionContourPx }),
+      () => ({ id: props.id, position: props.position, scale: props.scale, rotation: props.rotation, castShadow: props.castShadow, receiveShadow: props.receiveShadow, density: props.density, fontSize: props.fontSize, lineHeight: props.lineHeight, transparent: props.transparent, glyphPalette: props.glyphPalette, ambientIntensity: props.ambientIntensity, occlusionPriority: props.occlusionPriority, occlusionClaim: props.occlusionClaim, occlusionContourPx: props.occlusionContourPx }),
       () => {
         const mesh = meshRef.value;
         if (!mesh) return;
